feat : 정산 배치 스킵 로깅 및 실패/스킵 집계 보강
  - SkipListener 추가로 읽기/처리/쓰기 스킵 시 경고 로그를 남기고 스텝에 등록.
  - 배치 리스너에서 실패/스킵 카운트를 ExecutionContext로 전달하고 afterJob에서 합산해 complete/fail 기록 시 사용.
  - 셀러 전환 시 커서를 초기화하여 빈 페이지 연속 시 잘못된 lastCursor가 남지 않도록 수정.
  - 재시도/스킵 정책(재시도 3회, 지수 백오프, 4xx 스킵, 스킵 한도 50) 유지.

+++ modi/seller-service/src/main/java/com/coc/modi/seller/settlement/batch/SettlementBatchJobListener.java
@@ -1,7 +1,6 @@
 package com.coc.modi.seller.settlement.batch;
 
 import com.coc.modi.seller.settlement.application.SettlementBatchExecutionService;
-import com.coc.modi.seller.settlement.domain.SettlementBatchExecutionStatus;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.core.BatchStatus;
 import org.springframework.batch.core.JobExecution;
@@ -39,14 +38,14 @@ public class SettlementBatchJobListener implements JobExecutionListener {
             return;
         }
 
-        int readCount = jobExecution.getStepExecutions().stream()
-                .mapToInt(se -> se.getReadCount())
+        int readCount = (int) jobExecution.getStepExecutions().stream()
+                .mapToLong(se -> se.getReadCount())
                 .sum();
-        int writeCount = jobExecution.getStepExecutions().stream()
-                .mapToInt(se -> se.getWriteCount())
+        int writeCount = (int) jobExecution.getStepExecutions().stream()
+                .mapToLong(se -> se.getWriteCount())
                 .sum();
-        int processSkip = jobExecution.getStepExecutions().stream()
-                .mapToInt(se -> se.getProcessSkipCount() + se.getReadSkipCount() + se.getWriteSkipCount())
+        int processSkip = (int) jobExecution.getStepExecutions().stream()
+                .mapToLong(se -> se.getProcessSkipCount() + se.getReadSkipCount() + se.getWriteSkipCount())
                 .sum();
 
         BigDecimal totalAmount = jobExecution.getStepExecutions().stream()
@@ -65,8 +64,8 @@ public class SettlementBatchJobListener implements JobExecutionListener {
                 .mapToInt(se -> se.getFailureExceptions().size())
                 .sum();
 
-        int skipCount = jobExecution.getStepExecutions().stream()
-                .mapToInt(se -> se.getProcessSkipCount() + se.getReadSkipCount() + se.getWriteSkipCount())
+        int skipCount = (int) jobExecution.getStepExecutions().stream()
+                .mapToLong(se -> se.getProcessSkipCount() + se.getReadSkipCount() + se.getWriteSkipCount())
                 .sum();
 
         String lastCursor = jobExecution.getStepExecutions().stream()
